+++ test/locales/translations.test.js
@@ -176,7 +176,8 @@ const ignoredTransKeyForYourLoops = [
   'remote-monitoring',
   'no-settings-on-device-alert-message',
   'BGM',
-  'tooltip-empty-stat'
+  'tooltip-empty-stat',
+  'time-out-of-range'
 ]
 const ignoredTransKeyInYourLoopsFiles = [
   'yourloops|${s}',